Add files via upload
Well, this time the second test provided by the professor, after the corrections, shows that I got only one mistake. I decided to correct it latter on, because I'm interested in develop the second part of the activity that involves database. The final corrections, not the one that I mentioned previously, and the second test are in this commit.