Database implementation completion
This time I've made some major alterations on the project. As I started in the previous commit, I finished now the implementation of database, removed the classes and lines of code responsible for saving the information on the program, and now I'm working entirily with persistent storage. I tested with the three test classes, and everything is working.