Telegram bot
I've moved to the 3rd part of the project. Now the professor is requesting for a Telegram bot that interacts with the associates and provide them information about their fees and frequency. It works if the user asks directly for this things, like they was talking to someone (envolving natural language processing), or they can use the /consultar command that provides a ReplyKeyboardMarkup to the user with this two options, and when they select one of the options, the choosen information is returned as a message to the user.